requirements.txt
# DeepEval POC Requirements
deepeval>=0.21.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
